requirements.txt
fastapi[standard]
youtube-transcript-api==0.6.3
beautifulsoup4==4.13.3
fpdf==1.7.2
openai-whisper==20240930